requirements.txt
fastapi
uvicorn
pydantic
requests
pypdf
starlette
numpy
pandas
langchain
langchain-community
python-dotenv
google
google-generativeai
pinecone
unstructured